requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
black==25.1.0
certifi==2024.12.14
cfgv==3.4.0
click==8.1.8
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
fastapi==0.115.7
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.6
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
limits==4.0.1
loguru==0.7.3
mypy-extensions==1.0.0
nodeenv==1.9.1
openai==1.61.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
prometheus_client==0.21.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.34.0
virtualenv==20.29.1
wrapt==1.17.2
